requirements.txt
azure-functions==1.2.0
pysolr==3.8.1
pandas==1.0.1